Make Psalm happier
Let's add a bunch of types and annotations to remove most of the
INFO warnings from psalm. Most of these are very straightforward.

The most useful thing here is the templating of Some/None.

Note that i did not take to time to scrub `Either`. That is left as an
exercise to the library maintainer.