Option: Allow some impurity
Since we want to allow users to pass in function that are possibly non-pure
I am in debate around this because it opens up the library to mutation, based bugs.
If we force purity, then the utility of the library decreases.
EX: Option::some(false)->match(fn() => $this->passUnitTest(), fn() => $this->failUnitTest())